Move wishlist modals and share-button to /components/wishlist
