+++ core/components/wishlist/modals/delete.tsx
@@ -4,7 +4,7 @@ import { FormStatus } from '@/vibes/soul/form/form-status';
 import { Wishlist } from '@/vibes/soul/sections/wishlist-details';
 import { useModalForm } from '~/components/modal/modal-form-provider';
 
-import { deleteWishlistSchema } from '../../_actions/schema';
+import { deleteWishlistSchema } from '../../../app/[locale]/(default)/account/wishlists/_actions/schema';
 
 export const DeleteWishlistModal = ({ id, message }: Wishlist & { message: React.ReactNode }) => {
   const { state, form } = useModalForm(deleteWishlistSchema);
