+++ core/components/wishlist/modals/rename.tsx
@@ -8,7 +8,7 @@ import { Input } from '@/vibes/soul/form/input';
 import { Wishlist } from '@/vibes/soul/sections/wishlist-details';
 import { useModalForm } from '~/components/modal/modal-form-provider';
 
-import { renameWishlistSchema } from '../../_actions/schema';
+import { renameWishlistSchema } from '../../../app/[locale]/(default)/account/wishlists/_actions/schema';
 
 export const RenameWishlistModal = ({
   id,
